+++ src/app/(providers)/(root)/(mainpage)/selectedposts/page.tsx
@@ -1,10 +1,10 @@
 'use client';
 
-import { useEffect, useState, Suspense } from 'react';
-import { useRouter, useSearchParams } from 'next/navigation';
 import { createClient } from '@/utils/supabase/client';
 import Image from 'next/image';
 import Link from 'next/link';
+import { useRouter, useSearchParams } from 'next/navigation';
+import { Suspense, useEffect, useState } from 'react';
 
 const supabase = createClient();
 
@@ -31,10 +31,7 @@ const PostsList = () => {
     const startDate = searchParams.get('startDate');
     const endDate = searchParams.get('endDate');
 
-    let query = supabase
-      .from('posts')
-      .select('*')
-      .order('created_at', { ascending: false }); // 최신순으로 정렬
+    let query = supabase.from('posts').select('*').order('created_at', { ascending: false }); // 최신순으로 정렬
 
     if (selectedCity) {
       query = query.ilike('title', `%${selectedCity}%`);
@@ -70,38 +67,36 @@ const PostsList = () => {
   const formatPrice = (price: number) => {
     return new Intl.NumberFormat('en-US', {
       style: 'currency',
-      currency: 'USD',
+      currency: 'USD'
     }).format(price);
   };
 
   return (
     <div className="p-4">
       <button onClick={() => router.back()} className="mb-4">
-        ← 
+        ←
       </button>
 
-      <h2 className="text-xl font-bold mb-4">Related Posts</h2>
+      <h2 className="mb-4 text-xl font-bold">Related Posts</h2>
       {loading && <div>로딩 중...</div>}
       {error && <div>{error}</div>}
       {!loading && posts.length === 0 && <div>관련된 게시물이 없습니다.</div>}
       <ul>
         {posts.map((post) => (
-          <li key={post.id} className="mb-4 border p-2 rounded-md flex">
+          <li key={post.id} className="mb-4 flex rounded-md border p-2">
             <Link href={`/detail/${post.id}`} className="flex w-full">
               {post.image ? (
                 <Image src={post.image} alt={post.title} width={96} height={96} className="mr-2 w-24" />
               ) : (
-                <div className="mr-2 w-24 h-24 bg-gray-200 flex items-center justify-center">
-                  이미지 없음
-                </div>
+                <div className="mr-2 flex h-24 w-24 items-center justify-center bg-gray-200">이미지 없음</div>
               )}
               <div className="flex flex-col justify-between">
                 <div>
-                  <h3 className="font-bold text-xl">{post.title}</h3>
+                  <h3 className="text-xl font-bold">{post.title}</h3>
                   <p className="text-gray-500">{new Date(post.created_at).toLocaleDateString()}</p>
                   <p className="text-gray-700">{post.content}</p>
                 </div>
-                <div className="text-sm font-bold mt-2">{formatPrice(post.price)}</div>
+                <div className="mt-2 text-sm font-bold">{formatPrice(post.price)}</div>
               </div>
             </Link>
           </li>
